Moved app.tsx <main> render to Landing.tsx component
