+++ src/App.tsx
@@ -1,9 +1,9 @@
 import "./App.css";
 import Header from "./components/common/layout/header/Header";
 import Footer from "./components/common/layout/footer/Footer";
-import ProductList from "./components/product-list/ProductList";
-import TopSellers from "./components/top-sellers/TopSellers";
-import Testimonials from "./components/testimonials/testimonials";
+import ProductList from "./components/pages/landing/product-list/ProductList";
+import TopSellers from "./components/pages/landing/top-sellers/TopSellers";
+import Testimonials from "./components/pages/landing/testimonials/testimonials";
 
 function App() {
   return (
