+++ src/components/pages/landing/testimonials/sliderSettings.ts
